gateway.renproject.io: re-ordered storage providers
